pyproject.toml
[tool.poetry]
name = "graphql-sqlalchemy"
version = "0.6.1"
description = "Generate GraphQL Schemas from your SQLAlchemy models"
license = "MIT"
authors = ["Guido Rainuzzo <hi@guido.nyc>"]
readme = "README.md"
homepage = "https://github.com/gzzo/graphql-sqlalchemy"
repository = "https://github.com/gzzo/graphql-sqlalchemy"
keywords = ["graphql", "sqlalchemy"]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]

[tool.poetry.dependencies]
python = "<4,>=3.6.0"
graphql-core = ">=3.0.0,<4"
SQLAlchemy = "^1.3.0"
sphinx = {version = "==3.*,>=3.5.3", optional = true}
sphinx-rtd-theme = {version = "==0.*,>=0.5.1", optional = true}
pygments-graphql = {version = "==1.*,>=1.0.0", optional = true}
pygments-style-solarized = {version = "==0.*,>=0.1.1", optional = true}

[tool.poetry.dev-dependencies]
pytest = ">=3.4"
pytest-cov = ">=2.9.0"
pytest-describe = ">=1.0.0"
flake8 = ">=3.8.2"
black = ">=19.10b0"
mypy = ">=0.780"
codecov = ">=2.1.4"
tox = ">=3.15.1"
sqlalchemy-stubs = ">=0.4"
dephell = "==0.*,>=0.8.3"

[tool.poetry.extras]
docs = ["pygments-graphql", "pygments-style-solarized", "sphinx", "sphinx-rtd-theme"]


[tool.dephell.main]
versioning = "semver"
envs = ["main"]
from = {format = "poetry", path = "pyproject.toml"}

[tool.black]
target-version = ['py36', 'py37', 'py38']
line-length = 200

[build-system]
requires = ["poetry>=1,<2", "setuptools"]
build-backend = "poetry.masonry.api"

[pygments.lexers]
graphql = "pygments_graphql:GraphqlLexer"
